requirements.txt
numpy>=1.26.0
pandas>=2.2.3
setuptools>=79.0.1
minio>=7.2.15
scikit-learn>=1.6.1
pyyaml>=6.0.2
tensorflow>=2.19
comet-ml>=3.49.9
dvc>=3.59.1
dvc[s3]>=3.59.1
boto3>=1.35.62
botocore>=1.35.62
s3fs>=2025.3.2
awscli>=1.35.62
flask==3.1.0
psycopg2-binary==2.9.10
